fix(ui): resolve autoview display failures and standardize preview state - Unified redundant preview mode variables to prevent synchronization drift. - Corrected file list indexing for Autoview to account for absolute scroll offsets. - Integrated sorting into list construction to ensure consistent UI sourcing. - Resolved path handling and type conversion issues in file copy operations. - Modernized legacy type definitions and refined global window refresh logic.

+++ src/cmd/rmdir.c
@@ -5,8 +5,8 @@
  *
  ***************************************************************************/
 
+#include "ytree.h"
 #include "ytree_cmd.h"
-#include "ytree_defs.h"
 #include <errno.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -22,7 +22,7 @@ extern struct Volume *CurrentVolume;
 extern char *GetFileNamePath(FileEntry *file_entry, char *buffer);
 extern char *GetPath(DirEntry *dir_entry, char *buffer);
 extern int ScanSubTree(DirEntry *dir_entry, Statistic *s);
-extern int GetAvailBytes(LONGLONG *avail_bytes, Statistic *s);
+extern int GetAvailBytes(long long *avail_bytes, Statistic *s);
 
 #if defined(S_IFLNK)
 #define STAT_(a, b) lstat(a, b)
